♻️ relocate schema page components from `features/` to `components/`
