+++ frontend/apps/app/app/(app)/app/(with-project-and-branch)/projects/[projectId]/ref/[branchOrCommit]/schema/[...schemaFilePath]/page.tsx
@@ -1,5 +1,5 @@
 import type { PageProps } from '@/app/types'
-import { SchemaPage } from '@/features/schemas/pages/SchemaPage'
+import { SchemaPage } from '@/components/SchemaPage'
 import { branchOrCommitSchema } from '@/libs/routes'
 
 import * as v from 'valibot'
