+++ frontend/apps/app/components/ChatbotButton/components/ChatInput/index.ts
@@ -1,0 +1,1 @@
+export * from './ChatInput'
