+++ frontend/apps/app/components/ChatbotButton/components/ChatMessage/index.ts
@@ -1,0 +1,1 @@
+export * from './ChatMessage'
