+++ frontend/apps/app/components/ChatbotButton/components/ChatbotDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './ChatbotDialog'
