+++ frontend/apps/app/components/SchemaPage/SchemaPage.tsx
@@ -1,6 +1,5 @@
 import path from 'node:path'
 import { TabsContent, TabsRoot } from '@/components'
-import { ChatbotButton } from '@/features/schemas/components/Chatbot'
 import { createClient } from '@/libs/db/server'
 import { parse, setPrismWasmUrl } from '@liam-hq/db-structure/parser'
 import { getFileContent } from '@liam-hq/github'
@@ -8,6 +7,7 @@ import * as Sentry from '@sentry/nextjs'
 import { cookies } from 'next/headers'
 
 import type { ComponentProps, FC } from 'react'
+import { ChatbotButton } from '../ChatbotButton'
 import styles from './SchemaPage.module.css'
 import { ERDEditor } from './components/ERDEditor'
 import { OverrideEditor } from './components/OverrideEditor'
