+++ frontend/apps/app/components/SchemaPage/components/OverrideEditor/OverrideEditor.module.css
