+++ frontend/apps/app/components/SchemaPage/utils/safeApplySchemaOverride.ts
@@ -6,7 +6,7 @@ import {
 import { getFileContent } from '@liam-hq/github'
 import { safeParse } from 'valibot'
 import { parse as parseYaml } from 'yaml'
-import { SCHEMA_OVERRIDE_FILE_PATH } from '../../../constants'
+import { SCHEMA_OVERRIDE_FILE_PATH } from '../../../features/schemas/constants'
 
 export const safeApplySchemaOverride = async (
   repositoryFullName: string,
