--- frontend/apps/app/features/schemas/components/Chatbot/index.ts
@@ -1,4 +1,0 @@
-export * from './ChatbotButton'
-export * from './ChatbotDialog'
-export * from './ChatInput'
-export * from './ChatMessage'
